backend/requirements.txt
# Core dependencies
numpy>=1.19.5,<1.27.0
torch==2.0.1
torchvision==0.15.2

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.6

# Machine learning
scikit-learn==1.1.3
albumentations==1.3.1

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0

# Socket.IO dependencies
python-socketio[asyncio]==5.11.1
aiohttp==3.9.1

# Additional dependencies
opencv-python-headless==4.6.0.66
matplotlib==3.7.1
scipy>=1.5.0,<1.10.0
